Tsified lib/err-model
